add create transaction page for users, allow manager to edit user email and role
