+++ frontend/src/components/PromotionSelector.jsx
